+++ debug_toolbar/panels/templates/__init__.py
@@ -1,3 +1,3 @@
 from debug_toolbar.panels.templates.panel import TemplatesPanel
 
-__all__ = ["TemplatesPanel"]
+__all__ = [TemplatesPanel.panel_id]