+++ tests/test_integration.py
@@ -15,6 +15,12 @@ from django.test.utils import override_settings
 from debug_toolbar.forms import SignedDataForm
 from debug_toolbar.middleware import DebugToolbarMiddleware, show_toolbar
 from debug_toolbar.panels import Panel
+from debug_toolbar.panels.cache import CachePanel
+from debug_toolbar.panels.history import HistoryPanel
+from debug_toolbar.panels.request import RequestPanel
+from debug_toolbar.panels.sql import SQLPanel
+from debug_toolbar.panels.templates import TemplatesPanel
+from debug_toolbar.panels.versions import VersionsPanel
 from debug_toolbar.store import get_store
 from debug_toolbar.toolbar import DebugToolbar
 
@@ -98,7 +104,7 @@ class DebugToolbarTestCase(BaseTestCase):
     def _resolve_stats(self, path):
         # takes stats from Request panel
         self.request.path = path
-        panel = self.toolbar.get_panel_by_id("RequestPanel")
+        panel = self.toolbar.get_panel_by_id(RequestPanel.panel_id)
         response = panel.process_request(self.request)
         panel.generate_stats(self.request, response)
         return panel.get_stats()
@@ -149,9 +155,13 @@ class DebugToolbarTestCase(BaseTestCase):
         # may run earlier and cause fewer cache calls.
         cache.clear()
         response = self.client.get("/cached_view/")
-        self.assertEqual(len(response.toolbar.get_panel_by_id("CachePanel").calls), 3)
+        self.assertEqual(
+            len(response.toolbar.get_panel_by_id(CachePanel.panel_id).calls), 3
+        )
         response = self.client.get("/cached_view/")
-        self.assertEqual(len(response.toolbar.get_panel_by_id("CachePanel").calls), 2)
+        self.assertEqual(
+            len(response.toolbar.get_panel_by_id(CachePanel.panel_id).calls), 2
+        )
 
     @override_settings(ROOT_URLCONF="tests.urls_use_package_urls")
     def test_include_package_urls(self):
@@ -160,16 +170,24 @@ class DebugToolbarTestCase(BaseTestCase):
         # may run earlier and cause fewer cache calls.
         cache.clear()
         response = self.client.get("/cached_view/")
-        self.assertEqual(len(response.toolbar.get_panel_by_id("CachePanel").calls), 3)
+        self.assertEqual(
+            len(response.toolbar.get_panel_by_id(CachePanel.panel_id).calls), 3
+        )
         response = self.client.get("/cached_view/")
-        self.assertEqual(len(response.toolbar.get_panel_by_id("CachePanel").calls), 2)
+        self.assertEqual(
+            len(response.toolbar.get_panel_by_id(CachePanel.panel_id).calls), 2
+        )
 
     def test_low_level_cache_view(self):
         """Test cases when low level caching API is used within a request."""
         response = self.client.get("/cached_low_level_view/")
-        self.assertEqual(len(response.toolbar.get_panel_by_id("CachePanel").calls), 2)
+        self.assertEqual(
+            len(response.toolbar.get_panel_by_id(CachePanel.panel_id).calls), 2
+        )
         response = self.client.get("/cached_low_level_view/")
-        self.assertEqual(len(response.toolbar.get_panel_by_id("CachePanel").calls), 1)
+        self.assertEqual(
+            len(response.toolbar.get_panel_by_id(CachePanel.panel_id).calls), 1
+        )
 
     def test_cache_disable_instrumentation(self):
         """
@@ -181,7 +199,9 @@ class DebugToolbarTestCase(BaseTestCase):
         response = self.client.get("/execute_sql/")
         self.assertEqual(cache.get("UseCacheAfterToolbar.before"), 1)
         self.assertEqual(cache.get("UseCacheAfterToolbar.after"), 1)
-        self.assertEqual(len(response.toolbar.get_panel_by_id("CachePanel").calls), 0)
+        self.assertEqual(
+            len(response.toolbar.get_panel_by_id(CachePanel.panel_id).calls), 0
+        )
 
     def test_is_toolbar_request(self):
         self.request.path = "/__debug__/render_panel/"
@@ -254,8 +274,10 @@ class DebugToolbarIntegrationTestCase(IntegrationTestCase):
     def test_render_panel_checks_show_toolbar(self):
         url = "/__debug__/render_panel/"
         request_id = toolbar_request_id()
-        get_store().save_panel(request_id, "VersionsPanel", {"value": "Test data"})
-        data = {"request_id": request_id, "panel_id": "VersionsPanel"}
+        get_store().save_panel(
+            request_id, VersionsPanel.panel_id, {"value": "Test data"}
+        )
+        data = {"request_id": request_id, "panel_id": VersionsPanel.panel_id}
 
         response = self.client.get(url, data)
         self.assertEqual(response.status_code, 200)
@@ -283,7 +305,7 @@ class DebugToolbarIntegrationTestCase(IntegrationTestCase):
         self.assertEqual(len(request_ids), 1)
         toolbar = DebugToolbar.fetch(request_ids[0])
         self.assertEqual(
-            toolbar.get_panel_by_id("HistoryPanel").get_stats()["data"],
+            toolbar.get_panel_by_id(HistoryPanel.panel_id).get_stats()["data"],
             {"foo": "bar"},
         )
 
@@ -331,8 +353,8 @@ class DebugToolbarIntegrationTestCase(IntegrationTestCase):
         self.client.get("/execute_sql/")
         request_ids = list(get_store().request_ids())
         request_id = request_ids[-1]
-        toolbar = DebugToolbar.fetch(request_id, "SQLPanel")
-        panel = toolbar.get_panel_by_id("SQLPanel")
+        toolbar = DebugToolbar.fetch(request_id, SQLPanel.panel_id)
+        panel = toolbar.get_panel_by_id(SQLPanel.panel_id)
         djdt_query_id = panel.get_stats()["queries"][-1]["djdt_query_id"]
 
         url = "/__debug__/sql_select/"
@@ -361,8 +383,8 @@ class DebugToolbarIntegrationTestCase(IntegrationTestCase):
         self.client.get("/execute_sql/")
         request_ids = list(get_store().request_ids())
         request_id = request_ids[-1]
-        toolbar = DebugToolbar.fetch(request_id, "SQLPanel")
-        panel = toolbar.get_panel_by_id("SQLPanel")
+        toolbar = DebugToolbar.fetch(request_id, SQLPanel.panel_id)
+        panel = toolbar.get_panel_by_id(SQLPanel.panel_id)
         djdt_query_id = panel.get_stats()["queries"][-1]["djdt_query_id"]
 
         url = "/__debug__/sql_explain/"
@@ -394,8 +416,8 @@ class DebugToolbarIntegrationTestCase(IntegrationTestCase):
         self.client.get("/execute_json_sql/")
         request_ids = list(get_store().request_ids())
         request_id = request_ids[-1]
-        toolbar = DebugToolbar.fetch(request_id, "SQLPanel")
-        panel = toolbar.get_panel_by_id("SQLPanel")
+        toolbar = DebugToolbar.fetch(request_id, SQLPanel.panel_id)
+        panel = toolbar.get_panel_by_id(SQLPanel.panel_id)
         djdt_query_id = panel.get_stats()["queries"][-1]["djdt_query_id"]
 
         url = "/__debug__/sql_explain/"
@@ -423,8 +445,8 @@ class DebugToolbarIntegrationTestCase(IntegrationTestCase):
         self.client.get("/execute_sql/")
         request_ids = list(get_store().request_ids())
         request_id = request_ids[-1]
-        toolbar = DebugToolbar.fetch(request_id, "SQLPanel")
-        panel = toolbar.get_panel_by_id("SQLPanel")
+        toolbar = DebugToolbar.fetch(request_id, SQLPanel.panel_id)
+        panel = toolbar.get_panel_by_id(SQLPanel.panel_id)
         djdt_query_id = panel.get_stats()["queries"][-1]["djdt_query_id"]
 
         url = "/__debug__/sql_profile/"
@@ -532,7 +554,7 @@ class DebugToolbarIntegrationTestCase(IntegrationTestCase):
         request_id = el.attrib["data-request-id"]
         response = self.client.get(
             "/__debug__/render_panel/",
-            {"request_id": request_id, "panel_id": "TemplatesPanel"},
+            {"request_id": request_id, "panel_id": TemplatesPanel.panel_id},
         )
         self.assertEqual(response.status_code, 200)
         # The key None (without quotes) exists in the list of template
@@ -568,14 +590,14 @@ class DebugToolbarLiveTestCase(StaticLiveServerTestCase):
 
     def test_basic(self):
         self.get("/regular/basic/")
-        version_panel = self.selenium.find_element(By.ID, "VersionsPanel")
+        version_panel = self.selenium.find_element(By.ID, VersionsPanel.panel_id)
 
         # Versions panel isn't loaded
         with self.assertRaises(NoSuchElementException):
             version_panel.find_element(By.TAG_NAME, "table")
 
         # Click to show the versions panel
-        self.selenium.find_element(By.CLASS_NAME, "VersionsPanel").click()
+        self.selenium.find_element(By.CLASS_NAME, VersionsPanel.panel_id).click()
 
         # Version panel loads
         table = self.wait.until(
@@ -591,10 +613,10 @@ class DebugToolbarLiveTestCase(StaticLiveServerTestCase):
     )
     def test_basic_jinja(self):
         self.get("/regular_jinja/basic")
-        template_panel = self.selenium.find_element(By.ID, "TemplatesPanel")
+        template_panel = self.selenium.find_element(By.ID, TemplatesPanel.panel_id)
 
         # Click to show the template panel
-        self.selenium.find_element(By.CLASS_NAME, "TemplatesPanel").click()
+        self.selenium.find_element(By.CLASS_NAME, TemplatesPanel.panel_id).click()
 
         self.assertIn("Templates (2 rendered)", template_panel.text)
         self.assertIn("base.html", template_panel.text)
@@ -609,14 +631,14 @@ class DebugToolbarLiveTestCase(StaticLiveServerTestCase):
         self.get("/regular_jinja/basic")
         # Make a new request so the history panel has more than one option.
         self.get("/execute_sql/")
-        template_panel = self.selenium.find_element(By.ID, "HistoryPanel")
+        template_panel = self.selenium.find_element(By.ID, HistoryPanel.panel_id)
         # Record the current side panel of buttons for later comparison.
         previous_button_panel = self.selenium.find_element(
             By.ID, "djDebugPanelList"
         ).text
 
         # Click to show the history panel
-        self.selenium.find_element(By.CLASS_NAME, "HistoryPanel").click()
+        self.selenium.find_element(By.CLASS_NAME, HistoryPanel.panel_id).click()
         # Click to switch back to the jinja page view snapshot
         list(template_panel.find_elements(By.CSS_SELECTOR, "button"))[-1].click()
 
@@ -631,10 +653,10 @@ class DebugToolbarLiveTestCase(StaticLiveServerTestCase):
     @override_settings(DEBUG_TOOLBAR_CONFIG={"RESULTS_CACHE_SIZE": 0})
     def test_expired_store(self):
         self.get("/regular/basic/")
-        version_panel = self.selenium.find_element(By.ID, "VersionsPanel")
+        version_panel = self.selenium.find_element(By.ID, VersionsPanel.panel_id)
 
         # Click to show the version panel
-        self.selenium.find_element(By.CLASS_NAME, "VersionsPanel").click()
+        self.selenium.find_element(By.CLASS_NAME, VersionsPanel.panel_id).click()
 
         # Version panel doesn't loads
         error = self.wait.until(
@@ -662,10 +684,10 @@ class DebugToolbarLiveTestCase(StaticLiveServerTestCase):
     )
     def test_django_cached_template_loader(self):
         self.get("/regular/basic/")
-        version_panel = self.selenium.find_element(By.ID, "TemplatesPanel")
+        version_panel = self.selenium.find_element(By.ID, TemplatesPanel.panel_id)
 
         # Click to show the templates panel
-        self.selenium.find_element(By.CLASS_NAME, "TemplatesPanel").click()
+        self.selenium.find_element(By.CLASS_NAME, TemplatesPanel.panel_id).click()
 
         # Templates panel loads
         trigger = self.wait.until(
@@ -682,11 +704,11 @@ class DebugToolbarLiveTestCase(StaticLiveServerTestCase):
 
     def test_sql_action_and_go_back(self):
         self.get("/execute_sql/")
-        sql_panel = self.selenium.find_element(By.ID, "SQLPanel")
+        sql_panel = self.selenium.find_element(By.ID, SQLPanel.panel_id)
         debug_window = self.selenium.find_element(By.ID, "djDebugWindow")
 
         # Click to show the SQL panel
-        self.selenium.find_element(By.CLASS_NAME, "SQLPanel").click()
+        self.selenium.find_element(By.CLASS_NAME, SQLPanel.panel_id).click()
 
         # SQL panel loads
         button = self.wait.until(
@@ -709,7 +731,7 @@ class DebugToolbarLiveTestCase(StaticLiveServerTestCase):
     def test_displays_server_error(self):
         self.get("/regular/basic/")
         debug_window = self.selenium.find_element(By.ID, "djDebugWindow")
-        self.selenium.find_element(By.CLASS_NAME, "BuggyPanel").click()
+        self.selenium.find_element(By.CLASS_NAME, BuggyPanel.panel_id).click()
         self.wait.until(EC.visibility_of(debug_window))
         self.assertEqual(debug_window.text, "»\n500: Internal Server Error")
 
@@ -719,10 +741,10 @@ class DebugToolbarLiveTestCase(StaticLiveServerTestCase):
         assert hide_button.text == "Hide »"
 
         self.get("/execute_sql/")
-        sql_panel = self.selenium.find_element(By.ID, "SQLPanel")
+        sql_panel = self.selenium.find_element(By.ID, SQLPanel.panel_id)
 
         # Click to show the SQL panel
-        self.selenium.find_element(By.CLASS_NAME, "SQLPanel").click()
+        self.selenium.find_element(By.CLASS_NAME, SQLPanel.panel_id).click()
 
         table = self.wait.until(
             lambda selenium: sql_panel.find_element(By.TAG_NAME, "table")
@@ -737,10 +759,10 @@ class DebugToolbarLiveTestCase(StaticLiveServerTestCase):
         assert hide_button.text == "Esconder »"
 
         self.get("/execute_sql/")
-        sql_panel = self.selenium.find_element(By.ID, "SQLPanel")
+        sql_panel = self.selenium.find_element(By.ID, SQLPanel.panel_id)
 
         # Click to show the SQL panel
-        self.selenium.find_element(By.CLASS_NAME, "SQLPanel").click()
+        self.selenium.find_element(By.CLASS_NAME, SQLPanel.panel_id).click()
 
         table = self.wait.until(
             lambda selenium: sql_panel.find_element(By.TAG_NAME, "table")
@@ -756,10 +778,10 @@ class DebugToolbarLiveTestCase(StaticLiveServerTestCase):
         assert hide_button.text == "Hide »"
 
         self.get("/execute_sql/")
-        sql_panel = self.selenium.find_element(By.ID, "SQLPanel")
+        sql_panel = self.selenium.find_element(By.ID, SQLPanel.panel_id)
 
         # Click to show the SQL panel
-        self.selenium.find_element(By.CLASS_NAME, "SQLPanel").click()
+        self.selenium.find_element(By.CLASS_NAME, SQLPanel.panel_id).click()
 
         table = self.wait.until(
             lambda selenium: sql_panel.find_element(By.TAG_NAME, "table")